+++ alembic/versions/202506210037_hamali_labor_tracking.py
@@ -10,7 +10,7 @@ from sqlalchemy.dialects import postgresql
 from migration_utils import apply_updated_at_trigger, audit_columns, org_fk
 
 revision: str = "202506210037"
-down_revision: Union[str, None] = "202506210034"
+down_revision: Union[str, None] = "202506210036"
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None